Make Mailer implementation leaner
