+++ samples/play-easymail-usage/app/modules/CustomMailerBinder.java
@@ -1,8 +1,8 @@
 package modules;
 
 import com.feth.play.module.mail.Mailer;
+import com.feth.play.module.mail.Mailer.MailerFactory;
 import com.feth.play.module.mail.MailerImpl;
-import com.feth.play.module.mail.MailerImpl.MailerFactory;
 import com.google.inject.AbstractModule;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
