+++ samples/play-easymail-usage/build.sbt
@@ -6,7 +6,7 @@ version := "1.0-SNAPSHOT"
 
 libraryDependencies ++= Seq(
   // Comment the next line for local development:
-  "com.feth" %% "play-easymail" % "0.7.1-SNAPSHOT"
+  "com.feth" %% "play-easymail" % "0.8.1-SNAPSHOT"
 )
 //  Uncomment the next line for local development of the Play Easymail core:
 //lazy val playEasymail = project.in(file("modules/play-easymail")).enablePlugins(PlayJava)
